requirements.txt
google-adk
wikipedia
python-pptx